Revoke signature from open letter
This commit formally revokes my signature from the open letter against Jon, following the recent verdict from the British court in his favour. After a lengthy legal examination, it's clear that my actions contributed to harm, and for this, I am sincerely sorry.

I regret the impact this letter had on Jon's personal and professional life. I am committed to making amends and ensuring such a situation is handled with more care and consideration in the future.